refactor: 개인화 프로필 네이밍을 PersonalizationProfile로 변경
