+++ src/main/java/com/techfork/global/config/ElasticsearchCacheManager.java
@@ -5,7 +5,7 @@ import co.elastic.clients.elasticsearch._types.KnnSearch;
 import co.elastic.clients.json.JsonData;
 import com.techfork.domain.post.document.PostDocument;
 import com.techfork.domain.recommendation.config.RecommendationProperties;
-import com.techfork.domain.user.document.UserProfileDocument;
+import com.techfork.domain.user.document.PersonalizationProfileDocument;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.boot.ApplicationArguments;
@@ -103,7 +103,7 @@ public class ElasticsearchCacheManager implements ApplicationRunner {
                             .index(USER_PROFILES_INDEX)
                             .size(1)
                             .knn(List.of(createKnn("profileVector", dummyVector))),
-                    UserProfileDocument.class
+                    PersonalizationProfileDocument.class
             );
             log.debug("[ES Warmup] user_profiles kNN warmup OK");
         } catch (Exception e) {
